requirements.txt
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
google-generativeai==0.3.2

